fix eslint and component refactor
